Updated v0.1.4 anc refactor some helpers
