+++ x/migration/keeper/keeper.go
@@ -26,7 +26,6 @@ type (
 		sharedKeeper   types.SharedKeeper
 		appKeeper      types.ApplicationKeeper
 		supplierKeeper types.SupplierKeeper
-		authzKeeper    types.AuthzKeeper
 	}
 )
 
@@ -40,7 +39,6 @@ func NewKeeper(
 	sharedKeeper types.SharedKeeper,
 	appKeeper types.ApplicationKeeper,
 	supplierKeeper types.SupplierKeeper,
-	authzKeeper types.AuthzKeeper,
 ) Keeper {
 	if _, err := sdk.AccAddressFromBech32(authority); err != nil {
 		panic(fmt.Sprintf("invalid authority address: %s", authority))
@@ -56,7 +54,6 @@ func NewKeeper(
 		sharedKeeper:   sharedKeeper,
 		appKeeper:      appKeeper,
 		supplierKeeper: supplierKeeper,
-		authzKeeper:    authzKeeper,
 	}
 }
